Point2 holds integer coordinates

+++ libgdx/raytracer/core/src/codeguru/raytracer/Point.kt
@@ -1,5 +1,5 @@
 package codeguru.raytracer
 
-data class Point2(val x: Float, val y: Float)
+data class Point2(val x: Int, val y: Int)
 
 data class Point3(val x: Float, val y: Float, val z: Float)
